:truck: Rename : UserRepository 에서 관리하던 세션스토리지 메서드 별도로 분리 / deleteUser 메서드 추가
"구현 내용 :
- UserRepository 내에 존재했던 세션 관리 메서드들을 infrastructure 내에 SessionStorage 생성 후 내부로 이동
- 회원가입 시 UserInfo 저장하는 로직에서 뻑이 날 경우 User 데이터를 삭제하기 위한 deleteUser 메서드 추가"
