+++ src/main/java/domain/user/infrastructure/UserRepositoryImpl.java
@@ -1,12 +1,10 @@
 package domain.user.infrastructure;
 
-import static common.db.Database.SessionStorage;
 import static common.db.Database.Users;
 
 import domain.user.command.domain.User;
 import domain.user.command.domain.UserRepository;
 import java.util.Collection;
-import java.util.Map.Entry;
 import java.util.Optional;
 
 public class UserRepositoryImpl implements UserRepository {
@@ -26,20 +24,7 @@ public class UserRepositoryImpl implements UserRepository {
     }
 
     @Override
-    public void saveSession(String sessionId, String userId) {
-        SessionStorage().put(sessionId, userId);
-    }
-
-    @Override
-    public Optional<String> getSessionIdByUserId(String userId) {
-        return SessionStorage().entrySet().stream()
-            .filter(entry -> entry.getValue().equals(userId))
-            .map(Entry::getKey)
-            .findFirst();
-    }
-
-    @Override
-    public void removeSession(String sessionId) {
-        SessionStorage().remove(sessionId);
+    public void deleteUser(String userId) {
+        Users().remove(userId);
     }
 }
